requirements.txt
python>=3.10
python-kasa==0.6.2.1
python-dotenv==1.0.1
